Update README.md
"Wicked" (2025) is based on the Broadway Musical "Wicked: The Untold Story of the Witched of Oz" (2004), which is based on the book by Gregory Maguire "Wicked: The Life and Times of the Wicked Witch of the West". Which is also based on the 1939 film and 1900 book by L. Frank Baum "The Wizard of Oz".